requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
decorator==4.4.1
fake-useragent==0.1.11
idna==2.8
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
lxml==4.4.2
more-itertools==8.0.2
packaging==19.2
parse==1.13.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
py==1.8.1
pyee==6.0.0
Pygments==2.5.2
pyparsing==2.4.6
pyppeteer==0.0.25
pyquery==1.4.1
pytest==5.3.2
PyYAML==5.2
requests==2.22.0
requests-html==0.10.0
six==1.10.0
soupsieve==1.9.5
tqdm==4.41.0
traitlets==4.3.3
urllib3==1.25.7
w3lib==1.21.0
wcwidth==0.1.8
websockets==8.1
wykop-sdk==0.5.0
youtube-dl==2020.1.1
